Cleanup ListView (#366)
