+++ plugins/ui/src/js/src/elements/ListView.tsx
@@ -1,48 +1,18 @@
-import React, { ReactElement } from 'react';
 import { useSelector } from 'react-redux';
-import type { SelectionMode } from '@react-types/shared';
 import { isElementOfType } from '@deephaven/react-hooks';
 import { getSettings, RootState } from '@deephaven/redux';
-import {
-  ListView as DHListView,
-  ListViewProps as DHListViewProps,
-} from '@deephaven/components';
-import {
-  ListView as DHListViewJSApi,
-  ListViewProps as DHListViewJSApiProps,
-} from '@deephaven/jsapi-components';
-import {
-  SerializedListViewEventProps,
-  useListViewProps,
-} from './useListViewProps';
-import ObjectView, { ObjectViewProps } from './ObjectView';
+import { ListView as DHListView } from '@deephaven/components';
+import { ListView as DHListViewJSApi } from '@deephaven/jsapi-components';
+import { SerializedListViewProps, useListViewProps } from './useListViewProps';
+import ObjectView from './ObjectView';
 import useReExportedTable from './useReExportedTable';
 
-type WrappedDHListViewJSApiProps = Omit<DHListViewJSApiProps, 'table'> & {
-  children: ReactElement<ObjectViewProps>;
-};
-
-type WrappedDHListViewProps = Omit<DHListViewProps, 'selectionMode'> & {
-  // The spec specifies that selectionMode should be uppercase, but the Spectrum
-  // prop is lowercase. We'll accept either to keep things more flexible.s
-  selectionMode?: SelectionMode | Uppercase<SelectionMode>;
-};
-
-export type ListViewProps = (
-  | WrappedDHListViewProps
-  | WrappedDHListViewJSApiProps
-) &
-  SerializedListViewEventProps;
-
-function ListView({ children, selectionMode, ...props }: ListViewProps) {
+function ListView(props: SerializedListViewProps): JSX.Element | null {
   const settings = useSelector(getSettings<RootState>);
-  const listViewProps = useListViewProps(props);
+  const { children, ...listViewProps } = useListViewProps(props);
 
   const isObjectView = isElementOfType(children, ObjectView);
   const table = useReExportedTable(children);
-
-  const selectionModeLc = (selectionMode?.toLowerCase() ??
-    'none') as SelectionMode;
 
   if (isObjectView) {
     return (
@@ -50,7 +20,6 @@ function ListView({ children, selectionMode, ...props }: ListViewProps) {
         <DHListViewJSApi
           // eslint-disable-next-line react/jsx-props-no-spreading
           {...listViewProps}
-          selectionMode={selectionModeLc}
           table={table}
           settings={settings}
         />
@@ -62,7 +31,6 @@ function ListView({ children, selectionMode, ...props }: ListViewProps) {
     <DHListView
       // eslint-disable-next-line react/jsx-props-no-spreading
       {...listViewProps}
-      selectionMode={selectionModeLc}
     >
       {children}
     </DHListView>
